Legend: clean code - moved data folder in stories folder - responsive-chart-wrapper.js: deleted unused code - yarn.lock: added html-entities - stories: fixed new data location in files - data.test.js: fixed new data location
