+++ stories/bar.stories.js
@@ -2,9 +2,9 @@ import React, { useState } from 'react'
 import { storiesOf } from '@storybook/react'
 
 import BarChart from '../src/components/bar-chart/'
-import barChartData from '../src/shared/data/bar-chart-data'
-import scatterChartData from '../src/shared/data/scatter-chart-data'
-import lineChartData from '../src/shared/data/line-chart-data'
+import barChartData from './data/bar-chart-data'
+import scatterChartData from './data/scatter-chart-data'
+import lineChartData from './data/line-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
