+++ stories/data/bar-chart-data.js
