+++ stories/data/line-chart-data.js
