+++ stories/data/scatter-chart-data.js
