+++ stories/pie.stories.js
@@ -2,9 +2,9 @@ import React, { useState } from 'react'
 import { storiesOf } from '@storybook/react'
 
 import PieChart from '../src/components/pie-chart'
-import pieChartData from '../src/shared/data/pie-chart-data'
-import barChartData from '../src/shared/data/bar-chart-data'
-import lineChartData from '../src/shared/data/line-chart-data'
+import pieChartData from './data/pie-chart-data'
+import barChartData from './data/bar-chart-data'
+import lineChartData from './data/line-chart-data'
 import ResponsiveChartWrapper from './responsive-chart-wrapper'
